helper function created
