+++ constants/constants.cpp
@@ -1,0 +1,9 @@
+#include "constants.hpp"
+
+std::string decodedDate[51] = {"Jan","Feb","Mar","Apr","May","Jun", " ","\n",
+                               "Jul","Aug","Sep","Oct","Nov","Dec", " ","\n",
+                                "01", "02", "03", "04", "05", "06","07","\n",
+                                "08", "09", "10", "11", "12", "13","14","\n",
+                                "15", "16", "17", "18", "19", "20","21","\n",
+                                "22", "23", "24", "25", "26", "27","28","\n",
+                                "29", "30", "31"};
