+++ constants/constants.hpp
@@ -1,0 +1,7 @@
+#ifndef CONSTANTS_HPP_
+#define CONSTANTS_HPP_
+
+#include <string>
+extern std::string decodedDate[];
+
+#endif // !CONSTANTS_HPP_
