+++ helper_functions.hpp
@@ -1,0 +1,10 @@
+#ifndef HELPER_FUNCTION_HPP_
+#define HELPER_FUNCTION_HPP
+
+#include <string>
+#include "constants/constants.hpp"
+
+std::string decode_date(std::string boardRep);
+unsigned int date_to_index(std::string input);
+
+#endif // !HELPER_FUNCTION_HPP_
