+++ main.cpp
@@ -8,57 +8,13 @@
 #include <stdlib.h>
 
 #include "puzzle_piece/puzzle_piece.hpp"
-#include "helping_structure.hpp"
+#include "constants/constants.hpp"
+#include "helper_functions.hpp"
 
-std::string decodedDate[51] = {"Jan","Feb","Mar","Apr","May","Jun", " ","\n",
-                               "Jul","Aug","Sep","Oct","Nov","Dec", " ","\n",
-                                "01", "02", "03", "04", "05", "06","07","\n",
-                                "08", "09", "10", "11", "12", "13","14","\n",
-                                "15", "16", "17", "18", "19", "20","21","\n",
-                                "22", "23", "24", "25", "26", "27","28","\n",
-                                "29", "30", "31"};
 #define MAX_LOCATION 38
 #define DEBUG   0
 
-std::string decode_date(std::string boardRep)
-{
-    std::string date = "";
-    unsigned int loc=0;
-    loc = boardRep.rfind('.');
-    if(loc != std::string::npos)
-    {
-        date+=decodedDate[loc]+"-";
-        loc = boardRep.rfind('.',loc-1);
-        if(loc != std::string::npos)
-        {
-            date+=decodedDate[loc];
-        }
-    }
-    return date;
-}
 
-unsigned int date_to_index(std::string input)
-{
-    std::string *ptrfound;
-    unsigned int month = 12;
-    unsigned int day = 0;
-    if(input.length() == 6)
-    {
-        ptrfound = std::find(decodedDate,decodedDate+51,input.substr(3,3));
-        if(ptrfound < decodedDate+18)
-        {
-            month = ptrfound-decodedDate;
-            month = ((month/8)*6) + (month%8);
-        }
-        ptrfound = std::find(decodedDate,decodedDate+51,input.substr(0,2));
-        if(ptrfound < decodedDate+51)
-        {
-            day = ptrfound-decodedDate;
-            day = (((day/8)-2)*7) + (day%8);
-        }
-    }
-    return ((month)*31)+day;
-}
 
 class puzzle_board
 {
